Add coins (#5)
Adds the `Coin` class, which is an `Entity` that represents a
collectable coin. Once a `Robot` touches a `Coin`, it adds 1 to its coin
score. Coin scores are displayed on the left side of the viewport, and
they will be used as a tiebreaker in case of a stalemate. The `Arena`
spawns one `Coin` in a random location, and spawns a new one each time
it is collected. `Robot`s can now also detect the nearest `Coin`.