Improve pathfinding (#6)
Switch from a grid-based pathfinding system with rectilinear coordinates
to a vertex visibility pathfinding system.

To accomplish this, we create an extended hitbox for `Wall`s called
`pathfinding_hitbox`, which extends each side by `ROBOT_WIDTH + 2` (the
extra 2 pixels are arbitrary surplus). Then, we use hitbox vertices as
pathfinding nodes, with edges determined by visibility between nodes.

This works quite nicely, and it generates very quickly for a map.
However, it does incur some performance cost as checking for the nearest
visible node to an arbitrary point is expensive; we could optimize this
later via a `Quadtree`.